Modified the extension to accept real double
